# docs/swagger.json
{
    "swagger": "2.0",
    "info": {
        "description": "HIX-AI-2API",
        "title": "HIX-AI-2API",
        "contact": {},
        "version": "1.0.0"
    },
    "paths": {
        "/api/cookie": {
            "put": {
                "description": "保存COOKIE",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "COOKIE"
                ],
                "parameters": [
                    {
                        "type": "string",
                        "description": "Authorization BACKEND_SECRET",
                        "name": "Authorization",
                        "in": "header",
                        "required": true
                    },
                    {
                        "description": "COOKIE信息",
                        "name": "req",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/model.CookieSaveReq"
                        }
                    }
                ],
                "responses": {}
            }
        },
        "/api/cookie/all": {
            "get": {
                "description": "查询全量COOKIE",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "COOKIE"
                ],
                "parameters": [
                    {
                        "type": "string",
                        "description": "Authorization BACKEND_SECRET",
                        "name": "Authorization",
                        "in": "header",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "成功",
                        "schema": {
                            "allOf": [
                                {
                                    "$ref": "#/definitions/common.ResponseResult"
                                },
                                {
                                    "type": "object",
                                    "properties": {
                                        "data": {
                                            "type": "array",
                                            "items": {
                                                "$ref": "#/definitions/model.CookieResp"
                                            }
                                        }
                                    }
                                }
                            ]
                        }
                    }
                }
            }
        },
        "/api/cookie/credit/refresh": {
            "post": {
                "description": "同步更新全量COOKIE额度",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "COOKIE"
                ],
                "parameters": [
                    {
                        "type": "string",
                        "description": "Authorization BACKEND_SECRET",
                        "name": "Authorization",
                        "in": "header",
                        "required": true
                    }
                ],
                "responses": {}
            }
        },
        "/api/cookie/update": {
            "post": {
                "description": "更新COOKIE",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "COOKIE"
                ],
                "parameters": [
                    {
                        "type": "string",
                        "description": "Authorization BACKEND_SECRET",
                        "name": "Authorization",
                        "in": "header",
                        "required": true
                    },
                    {
                        "description": "COOKIE信息",
                        "name": "req",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/model.CookieUpdateReq"
                        }
                    }
                ],
                "responses": {}
            }
        },
        "/api/cookie/{id}": {
            "delete": {
                "description": "删除COOKIE",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "COOKIE"
                ],
                "parameters": [
                    {
                        "type": "string",
                        "description": "Authorization BACKEND_SECRET",
                        "name": "Authorization",
                        "in": "header",
                        "required": true
                    },
                    {
                        "type": "string",
                        "description": "COOKIE ID",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {}
            }
        },
...
    },
    "definitions": {
        "common.ResponseResult": {
            "type": "object",
            "properties": {
                "code": {
                    "type": "integer"
                },
                "data": {},
                "message": {
                    "type": "string"
                }
            }
        },
...
        "model.CookieResp": {
            "type": "object",
            "properties": {
                "cookie": {
                    "description": "Cookie cookie",
                    "type": "string"
                },
                "createTime": {
                    "description": "CreateTime 创建时间",
                    "type": "string"
                },
                "credit": {
                    "description": "Credit 额度",
                    "type": "integer"
                },
                "id": {
                    "description": "Id Id",
                    "type": "string"
                },
                "remark": {
                    "description": "Remark 备注",
                    "type": "string"
                }
            }
        },
        "model.CookieSaveReq": {
            "type": "object",
            "properties": {
                "cookie": {
                    "description": "Cookie cookie",
                    "type": "string"
                },
                "remark": {
                    "description": "Remark 备注",
                    "type": "string"
                }
            }
        },
        "model.CookieUpdateReq": {
            "type": "object",
            "properties": {
                "cookie": {
                    "description": "Cookie cookie",
                    "type": "string"
                },
                "id": {
                    "description": "Id Id",
                    "type": "string"
                },
                "remark": {
                    "description": "Remark 备注",
                    "type": "string"
                }
            }
        },
...
    }
}